Added test if data directory exists and create log of data of last download

+++ run_analysis.R
@@ -1,13 +1,23 @@
-#if(!file.exists("rawdata")){
-#     dir.create("rawdata")
-#}
+## run_analysis.R
+#
+#    Course Project for "Getting and Cleaning Data"
+#    May 2015
 
-DateDownloaded<-Sys.time()
+
+## ASSUMPTIONS
 fileUrl<-"https://d396qusza40orc.cloudfront.net/getdata%2Fprojectfiles%2FUCI%20HAR%20Dataset.zip"
+init_dir<-"UCI HAR Dataset"
 
-download.file(fileUrl,destfile="./smartphone.zip",method="curl")
+if(!file.exists(init_dir)){
+     # Only download data if it is missing    
+     DateDownloaded<-Sys.time()
 
-unzip("./smartphone.zip")
-file.remove("smartphone.zip")
-
-## Create Log file to record date downloaded.
+     download.file(fileUrl,destfile="./smartphone.zip",method="curl")
+     unzip("./smartphone.zip")
+     file.remove("smartphone.zip")
+     
+     ## Create Log file to record date downloaded.
+     f_con<-file("lastdownload.log",open="wt")
+     writeLines(paste("Date Downloaded:",DateDownloaded),con=f_con)
+     close(f_con)
+}
